activate surround module to work for “SDR” transforms created using the single Output Transform model
Note that this block is still largely a placeholder and it is expected that this module will be updated to support more dynamic ranges and surround levels (i.e. dark, dim, & normal). This allows for fuller experimentation with the parameterized Output Transform model to create transforms similar to those included in v1.0. Please use with caution.