change name of LMT and update header comments
Name changed based on feedback believing the current name could be misleading. Header comments were updated similarly. It was feared that labeled as such, the transform could be construed as a “miracle fix” for clipped highlights when in reality it serves a very specific case to reduce saturation in bright blues so that color-blocking or magenta hue shift artifacts are less likely to occur as a result of clipping.